+++ src/components/svgs/app-icon.tsx
@@ -1,0 +1,150 @@
+import * as React from "react";
+
+const AppIcon = React.forwardRef<SVGSVGElement, React.SVGProps<SVGSVGElement>>(function AppIcon(
+  props: Omit<React.SVGProps<SVGSVGElement>, "ref">,
+  ref: React.ForwardedRef<SVGSVGElement>,
+) {
+  return (
+    <svg
+      ref={ref}
+      height="512"
+      preserveAspectRatio="xMidYMid meet"
+      version="1.0"
+      viewBox="0 0 384 391.91998"
+      width="512"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      zoomAndPan="magnify"
+      {...props}
+    >
+      <defs>
+        <g />
+        <clipPath id="262caaae83">
+          <path
+            clipRule="nonzero"
+            d="M 19.285156 19.78125 L 364.714844 19.78125 L 364.714844 365.210938 L 19.285156 365.210938 Z M 19.285156 19.78125 "
+          />
+        </clipPath>
+        <clipPath id="39d1854b73">
+          <path
+            clipRule="nonzero"
+            d="M 192 19.78125 L 364.714844 192.496094 L 192 365.210938 L 19.285156 192.496094 Z M 192 19.78125 "
+          />
+        </clipPath>
+        <clipPath id="2a4b7fc5cb">
+          <path
+            clipRule="nonzero"
+            d="M 0.285156 0.78125 L 345.714844 0.78125 L 345.714844 346.210938 L 0.285156 346.210938 Z M 0.285156 0.78125 "
+          />
+        </clipPath>
+        <clipPath id="546555946b">
+          <path
+            clipRule="nonzero"
+            d="M 173 0.78125 L 345.714844 173.496094 L 173 346.210938 L 0.285156 173.496094 Z M 173 0.78125 "
+          />
+        </clipPath>
+        <clipPath id="d335859980">
+          <rect height="347" width="346" x="0" y="0" />
+        </clipPath>
+        <clipPath id="0a6635d2a4">
+          <path
+            clipRule="nonzero"
+            d="M 66.386719 66.886719 L 317.613281 66.886719 L 317.613281 318.109375 L 66.386719 318.109375 Z M 66.386719 66.886719 "
+          />
+        </clipPath>
+        <clipPath id="2fd99b7ffa">
+          <path
+            clipRule="nonzero"
+            d="M 192 66.886719 L 317.613281 192.496094 L 192 318.109375 L 66.386719 192.496094 Z M 192 66.886719 "
+          />
+        </clipPath>
+        <clipPath id="5e12954ff8">
+          <path
+            clipRule="nonzero"
+            d="M 0.386719 0.886719 L 251.613281 0.886719 L 251.613281 252.109375 L 0.386719 252.109375 Z M 0.386719 0.886719 "
+          />
+        </clipPath>
+        <clipPath id="4c77e77c6e">
+          <path
+            clipRule="nonzero"
+            d="M 126 0.886719 L 251.613281 126.496094 L 126 252.109375 L 0.386719 126.496094 Z M 126 0.886719 "
+          />
+        </clipPath>
+        <clipPath id="a076f2ac62">
+          <rect height="253" width="252" x="0" y="0" />
+        </clipPath>
+        <clipPath id="c42679ca16">
+          <path clipRule="nonzero" d="M 16 89 L 43 89 L 43 242.71875 L 16 242.71875 Z M 16 89 " />
+        </clipPath>
+        <clipPath id="9190d9eb73">
+          <rect height="243" width="50" x="0" y="0" />
+        </clipPath>
+      </defs>
+      <g clipPath="url(#262caaae83)">
+        <g clipPath="url(#39d1854b73)">
+          <g transform="matrix(1, 0, 0, 1, 19, 19)">
+            <g clipPath="url(#d335859980)">
+              <g clipPath="url(#2a4b7fc5cb)">
+                <g clipPath="url(#546555946b)">
+                  <path
+                    d="M 0.285156 0.78125 L 345.714844 0.78125 L 345.714844 346.210938 L 0.285156 346.210938 Z M 0.285156 0.78125 "
+                    fill="#000000"
+                    fillOpacity="1"
+                    fillRule="nonzero"
+                  />
+                </g>
+              </g>
+            </g>
+          </g>
+        </g>
+      </g>
+      <g clipPath="url(#0a6635d2a4)">
+        <g clipPath="url(#2fd99b7ffa)">
+          <g transform="matrix(1, 0, 0, 1, 66, 66)">
+            <g clipPath="url(#a076f2ac62)">
+              <g clipPath="url(#5e12954ff8)">
+                <g clipPath="url(#4c77e77c6e)">
+                  <path
+                    d="M 0.386719 0.886719 L 251.613281 0.886719 L 251.613281 252.109375 L 0.386719 252.109375 Z M 0.386719 0.886719 "
+                    fill="#f0f0ff"
+                    fillOpacity="1"
+                    fillRule="nonzero"
+                  />
+                </g>
+              </g>
+            </g>
+          </g>
+        </g>
+      </g>
+      <g transform="matrix(1, 0, 0, 1, 167, 28)">
+        <g clipPath="url(#9190d9eb73)">
+          <g fill="#be0078" fillOpacity="1">
+            <g transform="translate(-4.360953, 237.853512)">
+              <g>
+                <path d="M 10.890625 -152.5625 L 37.046875 -152.5625 L 37.046875 -52.3125 L 10.890625 -52.3125 Z M 10.890625 -26.15625 L 37.046875 -26.15625 L 37.046875 0 L 10.890625 0 Z M 10.890625 -26.15625 " />
+              </g>
+            </g>
+          </g>
+          <g clipPath="url(#c42679ca16)">
+            <g fill="#00fafa" fillOpacity="1">
+              <g transform="translate(5.921634, 242.218217)">
+                <g>
+                  <path d="M 10.890625 -152.5625 L 37.046875 -152.5625 L 37.046875 -52.3125 L 10.890625 -52.3125 Z M 10.890625 -26.15625 L 37.046875 -26.15625 L 37.046875 0 L 10.890625 0 Z M 10.890625 -26.15625 " />
+                </g>
+              </g>
+            </g>
+          </g>
+          <g fill="#41145f" fillOpacity="1">
+            <g transform="translate(0.780341, 240.035863)">
+              <g>
+                <path d="M 10.890625 -152.5625 L 37.046875 -152.5625 L 37.046875 -52.3125 L 10.890625 -52.3125 Z M 10.890625 -26.15625 L 37.046875 -26.15625 L 37.046875 0 L 10.890625 0 Z M 10.890625 -26.15625 " />
+              </g>
+            </g>
+          </g>
+        </g>
+      </g>
+    </svg>
+  );
+});
+
+export default AppIcon;