+++ src/components/svgs/index.tsx
@@ -1,3 +1,4 @@
+import AppIcon from "./app-icon";
 import EditSquare from "./edit-square";
 import Function from "./function";
 import ManageHistory from "./manage-history";
@@ -5,4 +6,4 @@ import NoteStack from "./note-stack";
 import TrailLength from "./trail-length";
 import Weight from "./weight";
 
-export { ManageHistory, EditSquare, Function, Weight, NoteStack, TrailLength };
+export { ManageHistory, EditSquare, Function, Weight, NoteStack, TrailLength, AppIcon };